% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ClearDesertSpringsCache}
\alias{ClearDesertSpringsCache}
\title{Clear cached data}
\usage{
ClearDesertSpringsCache(silent = FALSE)
}
\arguments{
\item{silent}{Silence feedback message?}
}
\value{
`TRUE` if cache was cleared, `FALSE` if no cache found
}
\description{
Clear cached data
}
