% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agol-utils.R
\name{FetchAGOLLayers}
\alias{FetchAGOLLayers}
\title{Read data from the Desert Springs AGOL feature layers.}
\usage{
FetchAGOLLayers(data_path, lookup_path, agol_username, agol_password)
}
\arguments{
\item{data_path}{URL to Desert Springs feature service on AGOL.}

\item{lookup_path}{URL to feature service on AGOL containing Desert Springs lookup tables.}

\item{agol_username}{Username of headless AGOL account with permissions to view the feature service.}

\item{agol_password}{Password for headless AGOL account.}
}
\value{
A list of tibbles
}
\description{
Read data from the Desert Springs AGOL feature layers.
}
