% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{FlowCategoriesAnnualHeatMap}
\alias{FlowCategoriesAnnualHeatMap}
\title{Summary heat map of flow categories for annual springs}
\usage{
FlowCategoriesAnnualHeatMap(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
ggplot heat map plot
}
\description{
Summary heat map of flow categories for annual springs
}
\examples{
\dontrun{
    
    FlowCategoriesAnnualHeatMap()
    FlowCategoriesAnnualHeatMap(site = "LAKE_P_GRA0058", field.season = "2019")
    FlowCategoriesAnnualHeatMap(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
    
}
}
