% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{FlowCategoriesDiscontinuous}
\alias{FlowCategoriesDiscontinuous}
\title{Summary table of flow categories for discontinuous springbrooks: dry, wet soil, <10 m, 10-50 m, >50 m}
\usage{
FlowCategoriesDiscontinuous(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Summary table of flow categories for discontinuous springbrooks: dry, wet soil, <10 m, 10-50 m, >50 m
}
\examples{
\dontrun{
    
    FlowCategoriesDiscontinuous()
    FlowCategoriesDiscontinuous(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
    
}
}
