% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{FlowCategoriesMap}
\alias{FlowCategoriesMap}
\title{Map of spring flow categories for latest field season at each park}
\usage{
FlowCategoriesMap(interactive, park, site, field.season)
}
\arguments{
\item{interactive}{Optional. Choose "yes" or "no." Yes will allow the user to toggle between field seasons of data. No will show only the latest field season of data. If no argument is entered, function will default to "no" for greater accessibility.}

\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
leaflet map
}
\description{
Map of spring flow categories for latest field season at each park
}
\examples{
\dontrun{
    
    FlowCategoriesMap(interactive = "no")
    FlowCategoriesMap(interactive = "yes", site = "LAKE_P_GRA0058", field.season = "2019")
    FlowCategoriesMap(interactive = "yes", park = c("DEVA", "MOJA"), field.season = c("2017", "2018", "2021"))
    
}
}
