% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{FlowCategoriesThreeYearHeatMap}
\alias{FlowCategoriesThreeYearHeatMap}
\title{Summary heat map of flow categories for 3-yr springs}
\usage{
FlowCategoriesThreeYearHeatMap(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
ggplot heat map plot
}
\description{
Summary heat map of flow categories for 3-yr springs
}
\examples{
\dontrun{
    
    FlowCategoriesThreeYearHeatMap()
    FlowCategoriesThreeYearHeatMap(site = "PARA_P_WHI0054", field.season = "2020")
    FlowCategoriesThreeYearHeatMap(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
    
}
}
