% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{FlowCategoriesThreeYearPlot}
\alias{FlowCategoriesThreeYearPlot}
\title{Summary bar plot of flow categories for 3-yr springs}
\usage{
FlowCategoriesThreeYearPlot(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
ggplot bar plot
}
\description{
Summary bar plot of flow categories for 3-yr springs
}
\examples{
\dontrun{
    
    FlowCategoriesThreeYearPlot()
    FlowCategoriesThreeYearPlot(park = c("DEVA", "MOJA"), field.season = c("2017", "2018", "2019"))
    
}
}
