% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbance-qc.R
\name{FlowModPlot}
\alias{FlowModPlot}
\title{Bar plot with percent of springs with active and historic flow modification}
\usage{
FlowModPlot(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
ggplot bar plot
}
\description{
Bar plot with percent of springs with active and historic flow modification
}
\examples{
\dontrun{
    
    FlowModPlot()
    FlowModPlot(park = c("DEVA", "JOTR"))
    
}
}
