% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbance-qc.R
\name{FlowModStatus}
\alias{FlowModStatus}
\title{List of springs with active or historical flow modification}
\usage{
FlowModStatus(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
List of springs with active or historical flow modification
}
\examples{
\dontrun{
    
    FlowModStatus()
    FlowModStatus(site = "DEVA_P_ARR0137", field.season = "2019")
    FlowModStatus(park = c("JOTR", "LAKE"), field.season = c("2017", "2018", "2021"))
    
}
}
