% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetRawData}
\alias{GetRawData}
\title{Raw data dump}
\usage{
GetRawData(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Spring code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A list of dataframes containing raw desert springs data.
}
\description{
Raw data dump
}
