% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetSiteName}
\alias{GetSiteName}
\title{Get the name of a site from the site code}
\usage{
GetSiteName(site.code)
}
\arguments{
\item{site.code}{Spring code to get the name for, e.g. "LAKE_P_HOR0042".}
}
\value{
The name of the site
}
\description{
Get the name of a site from the site code
}
