% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbance-qc.R
\name{HumanUseMap}
\alias{HumanUseMap}
\title{Map of human use observations}
\usage{
HumanUseMap(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
leaflet map
}
\description{
Map of human use observations
}
\examples{
\dontrun{
    
    HumanUseMap()
    HumanUseMap(site = "LAKE_P_DRI0002", field.season = "2019")
    HumanUseMap(park = c("DEVA", "MOJA"), field.season = c("2017", "2018", "2021"))
    
}
}
