% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation-qc.R
\name{InvasivePlants}
\alias{InvasivePlants}
\title{Table of tamarisk, fountain grass, rabbitsfoot grass, date palm, and fan palm (non-JOTR) observations}
\usage{
InvasivePlants(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Table of tamarisk, fountain grass, rabbitsfoot grass, date palm, and fan palm (non-JOTR) observations
}
\examples{
\dontrun{
    
    InvasivePlants()
    InvasivePlants(site = "LAKE_P_HOR0042", field.season = "2020")
    InvasivePlants(park = c("MOJA", "PARA"), field.season = c("2017", "2019", "2020"))
}
}
