% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation-qc.R
\name{InvasivePlantsMap}
\alias{InvasivePlantsMap}
\title{Map of tamarisk, fountain grass, rabbitsfoot grass, date palm, and fan palm (non-JOTR) observations}
\usage{
InvasivePlantsMap(interactive, park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
leaflet map
}
\description{
Map of tamarisk, fountain grass, rabbitsfoot grass, date palm, and fan palm (non-JOTR) observations
}
\examples{
\dontrun{
    InvasivePlantsMap()
    InvasivePlantsMap(site = "LAKE_P_GET0066", field.season = "2019")
    InvasivePlantsMap(park = c("MOJA", "PARA"), field.season = c("2017", "2019", "2020"))
}
}
