% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LabelBoxplotSampleSize}
\alias{LabelBoxplotSampleSize}
\title{Label sample size on boxplots}
\usage{
LabelBoxplotSampleSize(position)
}
\arguments{
\item{position}{Either a function (e.g. max, min, median) or a number specifying where on the y-axis the sample size label should appear.}
}
\value{
A dataframe with a SampleSize column as well as any grouping columns (Park, SiteCode, FieldSeason) that are present in data.
}
\description{
Label sample size on boxplots
}
