% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation-qc.R
\name{LifeformsPresence}
\alias{LifeformsPresence}
\title{Table with summary of life form presence and rank}
\usage{
LifeformsPresence(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Table with summary of life form presence and rank
}
\examples{
\dontrun{
    LifeformsPresence()
    LifeformsPresence(park = c("MOJA", "PARA"), field.season = c("2017", "2019", "2020"))
}
}
