% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbance-qc.R
\name{LivestockMap}
\alias{LivestockMap}
\title{Map of livestock disturbance observations}
\usage{
LivestockMap(interactive, park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
leaflet map
}
\description{
Map of livestock disturbance observations
}
\examples{
\dontrun{
    LivestockMap()
    LivestockMap(site = "MOJA_P_WIL0224", field.season = "2019")
    LivestockMap(park = c("DEVA", "MOJA"), field.season = c("2017", "2018", "2021"))
}
}
