% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbance-qc.R
\name{LivestockObservations}
\alias{LivestockObservations}
\title{Table with livestock observations}
\usage{
LivestockObservations(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Table with livestock observations
}
\examples{
\dontrun{
    
    LivestockObservations()
    LivestockObservations(site = "MOJA_P_WIL0224", field.season = "2019")
    LivestockObservations(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
    
}
}
