% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LoadDesertSprings}
\alias{LoadDesertSprings}
\title{Load raw data into package environment}
\usage{
LoadDesertSprings(
 
    data_path = "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/service_815059c20fd448628dc23441f7a7c473/FeatureServer",
  ...
)
}
\arguments{
\item{data_path}{A path or URL to the data. Accepted inputs:
* a URL to the AGOL feature service containing the data
* a folder containing the data in csv format
* a .zip file containing the data in csv format
* `"database"` (connect to the deprecated SQL server database)}

\item{...}{Additional arguments to OpenDatabaseConnection (ignored if `data_path != "database"`)}

\item{use_default_sql}{Use default SQL database? Ignored if `data_path != "database"`.}

\item{sql_drv}{Driver to use to connect to database. Ignored if `data_path != "database"`.}
}
\value{
Invisibly return a list containing all raw data
}
\description{
Run this function before you do anything else.
}
\examples{
\dontrun{
LoadDesertSprings()  # Read from AGOL
LoadDesertSprings("database")  # Read from SQL db
LoadDesertSprings("path/to/csv/folder")  # Read from folder of CSV's
LoadDesertSprings("path/to/zipped/csvs.zip")  # Read from zip file of CSV's
}

}
