% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-qc.R
\name{LocationMap}
\alias{LocationMap}
\title{Map of annual and 3-year desert springs monitoring locations}
\usage{
LocationMap(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
leaflet map
}
\description{
Map of annual and 3-year desert springs monitoring locations
}
\examples{
\dontrun{
    LocationMap()
    LocationMap(site = "LAKE_P_GET0066")
    LocationMap(park = c("DEVA", "MOJA"))
}
}
