% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation-qc.R
\name{MostCommonLifeformsPlot}
\alias{MostCommonLifeformsPlot}
\title{Bar plot showing the most common life form categories at springs by park}
\usage{
MostCommonLifeformsPlot(
  conn,
  path.to.data,
  park,
<<<<<<< HEAD
=======
  site,
>>>>>>> master
  field.season,
  data.source = "database"
)
}
\arguments{
\item{conn}{Database connection generated from call to \code{OpenDatabaseConnection()}. Ignored if \code{data.source} is \code{"local"}.}

\item{path.to.data}{The directory containing the csv data exports generated from \code{SaveDataToCsv()}. Ignored if \code{data.source} is \code{"database"}.}

\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{data.source}{Character string indicating whether to access data in the live desert springs database (\code{"database"}, default) or to use data saved locally (\code{"local"}). In order to access the most up-to-date data, it is recommended that you select \code{"database"} unless you are working offline or your code will be shared with someone who doesn't have access to the database.}
}
\value{
<<<<<<< HEAD

=======
ggplot bar plot
>>>>>>> master
}
\description{
Bar plot showing the most common life form categories at springs by park
}
<<<<<<< HEAD
=======
\examples{
\dontrun{
    conn <- OpenDatabaseConnection()
    MostCommonLifeformsPlot(conn)
    MostCommonLifeformsPlot(conn, park = c("MOJA", "PARA"), field.season = c("2017", "2019", "2020"))
    MostCommonLifeformsPlot(path.to.data = "path/to/data", data.source = "local")
    CloseDatabaseConnection(conn)
}
}
>>>>>>> master
