% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{qcContinuousLength}
\alias{qcContinuousLength}
\title{Continuous surface water length > discontinuous surface water length}
\usage{
qcContinuousLength(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Continuous surface water length > discontinuous surface water length
}
\examples{
\dontrun{
    qcContinuousLength()
    qcContinuousLength(site = "LAKE_P_GRA0058", field.season = "2019")
    qcContinuousLength(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
}
}
