% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbance-qc.R
\name{qcFlowModDiscrepancies}
\alias{qcFlowModDiscrepancies}
\title{List of springs that have been given different flow modification status in different field seasons}
\usage{
qcFlowModDiscrepancies(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
List of springs that have been given different flow modification status in different field seasons
}
\examples{
\dontrun{
    qcFlowModDiscrepancies()
    qcFlowModDiscrepancies(site = "DEVA_P_WAU0880")
    qcFlowModDiscrepancies(park = c("DEVA", "JOTR"))
}
}
