% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbance-qc.R
\name{qcFlowModFormatted}
\alias{qcFlowModFormatted}
\title{Intermediate step used to reformat flow modification types that can be read by other functions.}
\usage{
qcFlowModFormatted(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Intermediate step used to reformat flow modification types that can be read by other functions.
}
