% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbance-qc.R
\name{qcFlowModNoHumanUse}
\alias{qcFlowModNoHumanUse}
\title{Flow modification exists, but no Human Use disturbance recorded}
\usage{
qcFlowModNoHumanUse(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Flow modification exists, but no Human Use disturbance recorded
}
\examples{
\dontrun{
    qcFlowModNoHuman()
    qcFlowModNoHuman(site = "LAKE_P_WIL0061", field.season = "2016")
    qcFlowModNoHuman(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
}
}
