% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbance-qc.R
\name{qcFlowModTypes}
\alias{qcFlowModTypes}
\title{List of springs where no flow modification was observed but a modification type was selected or where flow modification was observed but no modification type was selected}
\usage{
qcFlowModTypes(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
List of springs where no flow modification was observed but a modification type was selected or where flow modification was observed but no modification type was selected
}
\examples{
\dontrun{
    qcFlowModTypes()
    qcFlowModTypes(site = "DEVA_P_WAU0880")
    qcFlowModTypes(park = c("MOJA", "LAKE"))
}
}
