% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality.R
\name{qcLocalDOCheck}
\alias{qcLocalDOCheck}
\title{Check dissolved oxygen calibrations for the use of non-local dissolved oxgygen percent.}
\usage{
qcLocalDOCheck(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Check dissolved oxygen calibrations for the use of non-local dissolved oxgygen percent.
}
\examples{
\dontrun{
    
    qcLocalDOCheck()
    qcLocalDOCheck(site = "PARA_P_COY0069")
    qcLocalDOCheck(park = c("LAKE", "PARA"), field.season = c("2017", "2019", "2020"))
}
}
