% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{qcMissingSensors}
\alias{qcMissingSensors}
\title{Sensors were retrieved but missing, or sensors were not retrieved and not missing}
\usage{
qcMissingSensors(park, deployment.field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{deployment.field.season}{Optional. Field season name to filter on, e.g. "2019".#'}
}
\value{
A tibble
}
\description{
Sensors were retrieved but missing, or sensors were not retrieved and not missing
}
\examples{
\dontrun{
    
    qcRepeatVisits()
    qcRepeatVisits(site = "LAKE_P_GET0066", field.season = "2019")
    qcRepeatVisits(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
    
}

}
