% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-qc.R
\name{qcNotSampled}
\alias{qcNotSampled}
\title{Return list of springs that were not sampled during a field season when they were intended to be monitored}
\usage{
qcNotSampled(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Return list of springs that were not sampled during a field season when they were intended to be monitored
}
\examples{
\dontrun{
    qcNotSampled()
    qcNotSampled(site = "LAKE_P_GET0066", field.season = "2019")
    qcNotSampled(park = "DEVA", field.season = c("2018", "2020", "2021"))
}
}
