% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbance-qc.R
\name{qcOverallDisturbance}
\alias{qcOverallDisturbance}
\title{Overall disturbance is less than any other disturbance category}
\usage{
qcOverallDisturbance(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Overall disturbance is less than any other disturbance category
}
\examples{
\dontrun{
    qcOverallDisturbance()
    qcOverallDisturbance(site = "LAKE_P_GET0066", field.season = "2019")
    qcOverallDisturbance(park = c("DEVA", "JOTR"), field.season = c("2017", "2020", "2021"))
}
}
