% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-qc.R
\name{qcPanelCheck}
\alias{qcPanelCheck}
\title{Return lists of site visits to springs that are not currently in an active sample frame or panel}
\usage{
qcPanelCheck(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Return lists of site visits to springs that are not currently in an active sample frame or panel
}
\examples{
\dontrun{
    qcPanelCheck()
    qcPanelCheck(park = "MOJA", field.season = c("2016", "2022", "2023"))
}
}
