% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{qcSensorDiscrepancies}
\alias{qcSensorDiscrepancies}
\title{Sensor was recovered but missing, or sensor was not recovered and not missing, or sensor was not recovered but downloaded}
\usage{
qcSensorDiscrepancies(park, deployment.field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{deployment.field.season}{Optional. Field season name to filter on, e.g. "2019".#'}
}
\value{
A tibble
}
\description{
Sensor was recovered but missing, or sensor was not recovered and not missing, or sensor was not recovered but downloaded
}
\examples{
\dontrun{
    qcRepeatVisits()
    qcRepeatVisits(site = "LAKE_P_GET0066", field.season = "2019")
    qcRepeatVisits(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
}
}
