% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{qcSensorHeatmap}
\alias{qcSensorHeatmap}
\title{Plot sensor retrieval results over time as a heatmap.}
\usage{
qcSensorHeatmap(park)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}
}
\value{
ggplot heatmap
}
\description{
Plot sensor retrieval results over time as a heatmap.
}
\examples{
\dontrun{
    qcSensorHeatmap()
    qcSensorHeatmap(park = c("JOTR", "MOJA"))
}
}
