% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{qcSensorsNotDeployed}
\alias{qcSensorsNotDeployed}
\title{Springs where a sensor was not deployed during a field season}
\usage{
qcSensorsNotDeployed(park, site, deployment.field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{deployment.field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Springs where a sensor was not deployed during a field season
}
\examples{
\dontrun{
    qcSensorsNotDeployed()
    qcSensorsNotDeployed(site = "MOJA_P_UNN0014")
    qcSensorsNotDeployed(park = c("DEVA", "JOTR"), deployment.field.season = c("2017", "2018", "2021"))
}
}
