% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality.R
\name{qcSpCondCalCheck}
\alias{qcSpCondCalCheck}
\title{Check for suspicious specific conductance calibrations}
\usage{
qcSpCondCalCheck(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Check for suspicious specific conductance calibrations
}
\examples{
\dontrun{
    qcUniqueIDMissing()
    qcUniqueIDMissing(site = "LAKE_P_GET0066")
    qcUniqueIDMissing(park = c("LAKE", "PARA"), field.season = c("2016", "2017", "2020"))
}
}
