% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{qcSpringDryWater}
\alias{qcSpringDryWater}
\title{Spring is dry, estimated discharge > 0 or volumetric discharge > 0 or springbrook dimension > 0}
\usage{
qcSpringDryWater(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Spring is dry, estimated discharge > 0 or volumetric discharge > 0 or springbrook dimension > 0
}
\examples{
\dontrun{
    
    qcSpringDryWater()
    qcSpringDryWater(site = "LAKE_P_GRA0058", field.season = "2019")
    qcSpringDryWater(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
    
}
}
