% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{qcSpringNotDryNoDischarge}
\alias{qcSpringNotDryNoDischarge}
\title{Spring is not dry, estimated discharge = 0 or volumetric discharge = 0}
\usage{
qcSpringNotDryNoDischarge(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Spring is not dry, estimated discharge = 0 or volumetric discharge = 0
}
\examples{
\dontrun{
    
    qcSpringNotDryNoDischarge()
    qcSpringNotDryNoDischarge(site = "LAKE_P_GRA0058", field.season = "2019")
    qcSpringNotDryNoDischarge(park = c("DEVA", "JOTR"), field.season = c("2017", "2019", "2020"))
    
}
}
