% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{qcSpringNotDryNoSpringbrook}
\alias{qcSpringNotDryNoSpringbrook}
\title{Spring is not dry, springbrook dimensions = 0}
\usage{
qcSpringNotDryNoSpringbrook(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Spring is not dry, springbrook dimensions = 0
}
\examples{
\dontrun{
    qcSpringNotDryNoSpringbrook()
    qcSpringNotDryNoSpringbrook(site = "DEVA_P_ARR0137", field.season = "2019")
    qcSpringNotDryNoSpringbrook(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
}
}
