% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-qc.R
\name{qcSpringTypeDiscrepancies}
\alias{qcSpringTypeDiscrepancies}
\title{Return list of springs that have been given different classifications}
\usage{
qcSpringTypeDiscrepancies(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Return list of springs that have been given different classifications
}
\examples{
\dontrun{
    qcSpringTypeDiscrepancies()
    qcSpringTypeDiscrepancies(site = "LAKE_P_GET0066")
    qcSpringTypeDiscrepancies(park = "DEVA", field.season = c("2018", "2021"))
}
}
