% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{qcUnknownSensorIDs}
\alias{qcUnknownSensorIDs}
\title{Sensors with unknown ID or serial number}
\usage{
qcUnknownSensorIDs(park, site, deployment.field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{deployment.field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Sensors with unknown ID or serial number
}
\examples{
\dontrun{
    qcRepeatVisits()
    qcRepeatVisits(site = "LAKE_P_GET0066", field.season = "2019")
    qcRepeatVisits(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
}
}
