% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation-qc.R
\name{qcVegDuplicates}
\alias{qcVegDuplicates}
\title{Return list of lifeform types that were duplicated during data entry}
\usage{
qcVegDuplicates(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Return list of lifeform types that were duplicated during data entry
}
\examples{
\dontrun{
    qcVegDuplicates()
    qcVegDuplicates(park = c("JOTR", "MOJA"), field.season = c("2016", "2018", "2021"))
}
}
