% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation-qc.R
\name{qcVegPresentNoLifeforms}
\alias{qcVegPresentNoLifeforms}
\title{Return list of visits with vegetation observed, no lifeform present}
\usage{
qcVegPresentNoLifeforms(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Return list of visits with vegetation observed, no lifeform present
}
\examples{
\dontrun{
    qcVegPresentNoLifeforms()
    qcVegPresentNoLifeforms(site = "LAKE_P_GET0066", field.season = "2019")
    qcVegPresentNoLifeforms(park = c("MOJA", "PARA"), field.season = c("2017", "2019", "2020"))
}
}
