% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-qc.R
\name{qcVisitDateTimelines}
\alias{qcVisitDateTimelines}
\title{Generate timeline of dates that each spring has been visited}
\usage{
qcVisitDateTimelines(
  conn,
  path.to.data,
  park,
  site,
  field.season,
  data.source = "database"
)
}
\arguments{
\item{conn}{Database connection generated from call to \code{OpenDatabaseConnection()}. Ignored if \code{data.source} is \code{"local"}.}

\item{path.to.data}{The directory containing the csv data exports generated from \code{SaveDataToCsv()}. Ignored if \code{data.source} is \code{"database"}.}

\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{data.source}{Character string indicating whether to access data in the live desert springs database (\code{"database"}, default) or to use data saved locally (\code{"local"}). In order to access the most up-to-date data, it is recommended that you select \code{"database"} unless you are working offline or your code will be shared with someone who doesn't have access to the database.}
}
\value{
<<<<<<< HEAD
A ggplot object
=======
ggplot timeline
>>>>>>> master
}
\description{
Generate timeline of dates that each spring has been visited
}
<<<<<<< HEAD
=======
\examples{
\dontrun{
    conn <- OpenDatabaseConnection()
    qcVisitDateTimelines(conn)
    qcVisitDateTimelines(conn, site = "LAKE_P_GET0066", field.season = "2019")
    qcVisitDateTimelines(conn, park = "DEVA", field.season = c("2018", "2020", "2021"))
    qcVisitDateTimelines(path.to.data = "path/to/data", data.source = "local")
    CloseDatabaseConnection(conn)
}
}
>>>>>>> master
