% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-qc.R
\name{qcVisitsByDate}
\alias{qcVisitsByDate}
\title{Return list of springs that have been visited on each date}
\usage{
qcVisitsByDate(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Return list of springs that have been visited on each date
}
\examples{
\dontrun{
    qcVisitsByDate()
    qcVisitsByDate(site = "LAKE_P_GET0066", field.season = "2019")
    qcVisitsByDate(park = "DEVA", field.season = c("2018", "2020", "2021"))
}
}
