% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{qcVolumetricFillEvents}
\alias{qcVolumetricFillEvents}
\title{Volumetric method was used, but there are fewer than five fill times}
\usage{
qcVolumetricFillEvents(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Volumetric method was used, but there are fewer than five fill times
}
\examples{
\dontrun{
    
    qcVolumetricFillEvents()
    qcVolumetricFillEvents(site = "MOJA_P_MAR0147", field.season = "2019")
    qcVolumetricFillEvents(park = c("DEVA", "MOJA"), field.season = c("2017", "2019", "2021"))
    
}
}
