% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{qcVolumetricMissing}
\alias{qcVolumetricMissing}
\title{Volumetric method was used, but there is no container volume, percent of flow, or fill time recorded}
\usage{
qcVolumetricMissing(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Volumetric method was used, but there is no container volume, percent of flow, or fill time recorded
}
\examples{
\dontrun{
    
    qcVolumetricMissing()
    qcVolumetricMissing(site = "LAKE_P_GRA0058", field.season = "2019")
    qcVolumetricMissing(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
    
}
}
