% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{qcVolumetricTimes}
\alias{qcVolumetricTimes}
\title{Volumetric method was used, but the median fill time is less than 5 seconds}
\usage{
qcVolumetricTimes(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Volumetric method was used, but the median fill time is less than 5 seconds
}
\examples{
\dontrun{
    qcVolumetricTimes()
    qcVolumetricTimes(site = "LAKE_P_GRA0058", field.season = "2019")
    qcVolumetricTimes(park = c("DEVA", "MOJA"), field.season = c("2017", "2018", "2019"))
}
}
