% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildlife-qc.R
\name{qcWildlifeDuplicates}
\alias{qcWildlifeDuplicates}
\title{Return list of wildlife types that were duplicated during data entry}
\usage{
qcWildlifeDuplicates(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Return list of wildlife types that were duplicated during data entry
}
\examples{
\dontrun{
    qcWildlifeDuplicates()
    qcWildlifeDuplicates(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
}
}
