% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildlife-qc.R
\name{qcWildlifeObservedNoEvidence}
\alias{qcWildlifeObservedNoEvidence}
\title{Wildlife observed and wildlife type specified, no evidence recorded}
\usage{
qcWildlifeObservedNoEvidence(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Wildlife observed and wildlife type specified, no evidence recorded
}
\examples{
\dontrun{
    qcWildlifeObservedNoEvidence()
    qcWildlifeObservedNoEvidence(site = "LAKE_P_GET0066", field.season = "2019")
    qcWildlifeObservedNoEvidence(park = c("DEVA", "JOTR"), field.season = c("2017", "2020", "2021"))
}
}
