% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality.R
\name{QcWqCleaned}
\alias{QcWqCleaned}
\title{Clean water quality data and format for plotting. Intermediate step.}
\usage{
QcWqCleaned(
  conn,
  path.to.data,
  park,
  site,
  field.season,
  data.source = "database"
)
}
\arguments{
\item{conn}{}

\item{path.to.data}{}

\item{park}{}

\item{site}{}

\item{field.season}{}

\item{data.source}{}
}
\value{

}
\description{
Clean water quality data and format for plotting. Intermediate step.
}
