% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality.R
\name{qcWqFlags}
\alias{qcWqFlags}
\title{Compile list of water quality values that have data quality flags.}
\usage{
qcWqFlags(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble with columns for Park, FieldSeason, SiteCode, VisitDate, Parameter, Units, Median, Flag, and FlagNote.
}
\description{
Compile list of water quality values that have data quality flags.
}
\examples{
\dontrun{
    qcWqFlags()
    qcWqFlags(site = "LAKE_P_GET0066", field.season = "2019")
    qcWqFlags(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
}
}
