% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality.R
\name{qcWqSanity}
\alias{qcWqSanity}
\title{Perform sanity check and compile list of potentially incorrect or outlier water quality values.}
\usage{
qcWqSanity(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble with columns for Park, FieldSeason, SiteCode, VisitDate, Parameter, Units, Median, Flag, and FlagNote.
}
\description{
Perform sanity check and compile list of potentially incorrect or outlier water quality values.
}
\examples{
\dontrun{
    qcWqSanity()
    qcWqSanity(site = "LAKE_P_HOT0065", field.season = "2019")
    qcWqSanity(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
}
}
