% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReadSqlDatabase}
\alias{ReadSqlDatabase}
\title{Read data from the Desert Springs SQL database}
\usage{
ReadSqlDatabase(...)
}
\arguments{
\item{...}{Optional arguments to be passed to `OpenDatabaseConnection()`}
}
\value{
A list of tibbles
}
\description{
Read data from the Desert Springs SQL database
}
