% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{SpringDischarge}
\alias{SpringDischarge}
\title{Join flow condition, estimated discharge, and volumetric discharge data into one table}
\usage{
SpringDischarge(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Join flow condition, estimated discharge, and volumetric discharge data into one table
}
\examples{
\dontrun{
    SpringDischarge()
    SpringDischarge(site = "LAKE_P_GRA0058", field.season = "2019")
    SpringDischarge(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
}
}
