% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildlife-qc.R
\name{UngulatesEvidence}
\alias{UngulatesEvidence}
\title{Table of springs with evidence of ungulate (sheep and deer) activity}
\usage{
UngulatesEvidence(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Table of springs with evidence of ungulate (sheep and deer) activity
}
\examples{
\dontrun{
    
    UngulatesEvidence()
    UngulatesEvidence(site = "LAKE_P_COR0023", field.season = "2020")
    UngulatesEvidence(park = c("DEVA", "JOTR"), field.season = c("2017", "2020", "2021"))
}
}
