% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildlife-qc.R
\name{UngulatesMap}
\alias{UngulatesMap}
\title{Map of springs with evidence of ungulate (sheep and deer) activity}
\usage{
UngulatesMap(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
leaflet map
}
\description{
Map of springs with evidence of ungulate (sheep and deer) activity
}
\examples{
\dontrun{
    
    UngulatesMap()
    UngulatesMap(site = "LAKE_P_COR0023")
    UngulatesMap(park = c("DEVA", "MOJA"), field.season = c("2019", "2021"))
}
}
