% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agol-utils.R
\name{unk_sensors}
\alias{unk_sensors}
\title{View sensors with unknown ID/serial number}
\usage{
unk_sensors()
}
\value{
Dataframe of sensors omitted from data due to missing serial numbers
}
\description{
View sensors with unknown ID/serial number
}
