% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{VolumetricMedian}
\alias{VolumetricMedian}
\title{Calculate median discharge and count of fill times for volumetric method}
\usage{
VolumetricMedian(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Calculate median discharge and count of fill times for volumetric method
}
\examples{
\dontrun{
    VolumetricMedian()
    VolumetricMedian(site = "LAKE_P_GRA0058", field.season = "2019")
    VolumetricMedian(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
}
}
