% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality.R
\name{WqMapPH}
\alias{WqMapPH}
\title{Map of pH at springs with surface water}
\usage{
WqMapPH(park, site, field.season = "database")
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
leaflet map
}
\description{
Map of pH at springs with surface water
}
\examples{
\dontrun{
    WqMapPH()
    WqMapPH(site = "LAKE_P_GET0066", field.season = "2019")
    WqMapPH(park = c("MOJA", "PARA"), field.season = c("2017", "2019", "2020"))
}
}
