% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality.R
\name{WqMapSpCond}
\alias{WqMapSpCond}
\title{Map of specific conductance at springs with surface water}
\usage{
WqMapSpCond(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
leaflet map
}
\description{
Map of specific conductance at springs with surface water
}
\examples{
\dontrun{
    WqMapSpCond()
    WqMapSpCond(site = "LAKE_P_GET0066", field.season = "2019")
    WqMapSpCond(park = c("MOJA", "PARA"), field.season = c("2017", "2019", "2020"))
}
}
