% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality.R
\name{WqMedian}
\alias{WqMedian}
\title{Calculate median values for each water quality parameter for each site visit.}
\usage{
WqMedian(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble with columns for park, field season, site code, visit date, and the median values, flags, and flag notes for temperature, specific conductance, pH, and dissolved oxygen.
}
\description{
Calculate median values for each water quality parameter for each site visit.
}
\examples{
\dontrun{
    
    WqMedian()
    WqMedian(site = "LAKE_P_GET0066", field.season = "2019")
    WqMedian(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
}
}
