% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality.R
\name{WqPlotGrid}
\alias{WqPlotGrid}
\title{Generate grid of box plots for core water quality parameters for each park and year. Includes annual and 3Yr springs only.}
\usage{
WqPlotGrid(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Grid of ggplot box plots
}
\description{
Generate grid of box plots for core water quality parameters for each park and year. Includes annual and 3Yr springs only.
}
\examples{
\dontrun{
    
    WqPlotGrid()
    WqPlotGrid(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
    WqPlotGrid(path.to.data = "path/to/data" = "local")
    CloseDatabaseection()
}
}
