% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality.R
\name{WqPlotSpCondmS}
\alias{WqPlotSpCondmS}
\title{Generate box plots for specific conductance for each park and year in units of mS/cm. Includes annual and 3Yr springs only.}
\usage{
WqPlotSpCondmS(park, site, field.season, include.title = FALSE)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{include.title}{Include plot title? Defaults to TRUE}
}
\value{
Box plots of specific conductance data for each park and field season.
}
\description{
Generate box plots for specific conductance for each park and year in units of mS/cm. Includes annual and 3Yr springs only.
}
