% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality.R
\name{WqStats}
\alias{WqStats}
\title{Calculate quartile values for each water quality parameter for each park and year. Includes annual and 3Yr springs only.}
\usage{
WqStats(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble with columns for Park; FieldSeason; Parameter; Units; and 0%, 25%, 50%, 75%, and 100% quantiles.
}
\description{
Calculate quartile values for each water quality parameter for each park and year. Includes annual and 3Yr springs only.
}
\examples{
\dontrun{
    WqStats()
    WqStats(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
}
}
