% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agol-utils.R
\name{WrangleAGOLData}
\alias{WrangleAGOLData}
\title{Wrangle AGOL data into a set of dataframes structured for use in this package.}
\usage{
WrangleAGOLData(agol_layers)
}
\arguments{
\item{agol_layers}{The list of tibbles returned by `FetchAGOLLayers()`}
}
\value{
A list of tibbles
}
\description{
Wrangle AGOL data into a set of dataframes structured for use in this package.
}
