% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LoadDesertSprings}
\alias{LoadDesertSprings}
\title{Load raw data into package environment}
\usage{
LoadDesertSprings(
  data_path = c(main_db =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_DS_SpringVisit/FeatureServer",
    lookup_db =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_Lookup_Database/FeatureServer",
    sites_db =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_DS_Sites_Master/FeatureServer",
    calibration_db =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_Calibration_Database/FeatureServer"),
  use_default_sql = FALSE,
  sql_drv = odbc::odbc(),
  agol_username = "mojn_hydro",
  agol_password =
    rstudioapi::askForPassword(paste("Please enter the password for AGOL account",
    agol_username)),
  ...
)
}
\arguments{
\item{data_path}{A path or URL to the data. Accepted inputs:
* 4 URLs to the AGOL feature services containing the data (main_db, lookup_db, sites_db, and calibration_db)
* a folder containing the data in csv format
* a .zip file containing the data in csv format
* `"database"` (connect to the deprecated SQL server database)}

\item{use_default_sql}{Use default SQL database? Ignored if `data_path != "database"`.}

\item{sql_drv}{Driver to use to connect to database. Ignored if `data_path != "database"`.}

\item{...}{Additional arguments to OpenDatabaseConnection (ignored if `data_path != "database"`)}
}
\value{
Invisibly return a list containing all raw data
}
\description{
Run this function before you do anything else.
}
\examples{
\dontrun{
LoadDesertSprings()  # Read from AGOL
LoadDesertSprings("database")  # Read from SQL db
LoadDesertSprings("path/to/csv/folder")  # Read from folder of CSV's
LoadDesertSprings("path/to/zipped/csvs.zip")  # Read from zip file of CSV's
}

}
