% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation-qc.R
\name{MostCommonLifeformsPlot}
\alias{MostCommonLifeformsPlot}
\title{Bar plot showing the most common life form categories at springs by park}
\usage{
MostCommonLifeformsPlot(park, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
ggplot bar plot
}
\description{
Bar plot showing the most common life form categories at springs by park
}
\examples{
\dontrun{
    MostCommonLifeformsPlot()
    MostCommonLifeformsPlot(park = c("MOJA", "PARA"), field.season = c("2017", "2019", "2020"))
}
}
