% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-qc.R
\name{qcCompleteness}
\alias{qcCompleteness}
\title{Calculate completeness (\% of planned sites visited)}
\usage{
qcCompleteness(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble with columns for park, field season, sample frame (i.e., annual, 3Yr), monitoring status (i.e., sampled), count of springs monitored, and percent of springs monitored.
}
\description{
Calculate completeness (\% of planned sites visited)
}
\examples{
\dontrun{
    qcCompleteness()
    qcCompleteness(site = "LAKE_P_GET0066", field.season = "2019")
    qcCompleteness(park = "DEVA", field.season = c("2018", "2020", "2021"))
}
}
