% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-qc.R
\name{qcCompletenessPlot}
\alias{qcCompletenessPlot}
\title{Generate stacked bar plot for completeness (\% of planned sites visited).}
\usage{
qcCompletenessPlot(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
ggplot stacked bar plot
}
\description{
Generate stacked bar plot for completeness (\% of planned sites visited).
}
\examples{
\dontrun{
    qcCompletenessPlot()
    qcCompletenessPlot(park = "DEVA", field.season = c("2018", "2020", "2021"))
}
}
