% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{qcDischargeMissing}
\alias{qcDischargeMissing}
\title{Volumetric or estimated discharge data are missing}
\usage{
qcDischargeMissing(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble
}
\description{
Volumetric or estimated discharge data are missing
}
\examples{
\dontrun{
    
    qcDischargeMissing()
    qcDischargeMissing(site = "LAKE_P_GET0066", field.season = "2016")
    qcDischargeMissing(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
    
}
}
