% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{qcSensorSummary}
\alias{qcSensorSummary}
\title{Summarize sensor retrieval and download attempts by park and season}
\usage{
qcSensorSummary(park, deployment.field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{deployment.field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
Tibble with columns for park, deployment field season, number of sensors deployed, number of sensors for which retrieval was not attempted, number of sensors for which retrieval was attempted, number of sensors actually retrieved, number of sensors actually downloaded
}
\description{
Summarize sensor retrieval and download attempts by park and season
}
\examples{
\dontrun{
    
    qcSensorSummary()
    qcSensorSummary(park = "DEVA", deployment.field.season = c("2018", "2020", "2021"))
    
}
}
