% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReadAGOL}
\alias{ReadAGOL}
\title{Read data from the Desert Springs AGOL feature layer.}
\usage{
ReadAGOL(
  data_path = c(main_db =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_DS_SpringVisit/FeatureServer",
    lookup_db =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_Lookup_Database/FeatureServer",
    sites_db =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_DS_Sites_Master/FeatureServer",
    calibration_db =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_Calibration_Database/FeatureServer"),
  agol_username = "mojn_hydro",
  agol_password =
    rstudioapi::askForPassword(paste("Please enter the password for AGOL account",
    agol_username))
)
}
\arguments{
\item{data_path}{URL to Desert Springs feature service on AGOL.}

\item{agol_username}{Username of headless AGOL account with permissions to view the feature service.}

\item{agol_password}{Password for headless AGOL account.}
}
\value{
A list of tibbles
}
\description{
Read data from the Desert Springs AGOL feature layer.
}
