% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{SpringbrookLengthsAnnualPlot}
\alias{SpringbrookLengthsAnnualPlot}
\title{Box plot of springbrook lengths for annual springs at each park and field season}
\usage{
SpringbrookLengthsAnnualPlot(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
ggplot box plot
}
\description{
Box plot of springbrook lengths for annual springs at each park and field season
}
\examples{
\dontrun{
    SpringbrookLengthsAnnualPlot()
    SpringbrookLengthsAnnualPlot(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
}
}
