% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge-qc.R
\name{SpringbrookLengthsThreeYearPlot}
\alias{SpringbrookLengthsThreeYearPlot}
\title{Box plot of springbrook lengths for three-year springs at each park and field season}
\usage{
SpringbrookLengthsThreeYearPlot(
  conn,
  path.to.data,
  park,
  site,
  field.season,
  data.source = "database"
)
}
\arguments{
\item{conn}{}

\item{path.to.data}{}

\item{park}{}

\item{site}{}

\item{field.season}{}

\item{data.source}{}
}
\value{

}
\description{
Box plot of springbrook lengths for three-year springs at each park and field season
}
