% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality.R
\name{WqPlotDOmgL}
\alias{WqPlotDOmgL}
\title{Generate box plots for concentration dissolved oxygen for each park and year. Includes annual and 3Yr springs only.}
\usage{
WqPlotDOmgL(park, site, field.season, include.title = FALSE)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{include.title}{}
}
\value{
ggplot box plots
}
\description{
Generate box plots for concentration dissolved oxygen for each park and year. Includes annual and 3Yr springs only.
}
\examples{
\dontrun{
    
    WqPlotDOmgL()
    WqPlotDOmgL(park = c("DEVA", "JOTR"), field.season = c("2017", "2018", "2021"))
}
}
